requirements.txt
numpy
flask
ipykernel
opencv-python
ultralytics
supervision
mysql-connector-python
datetime

